// src/api/model/policyResponse.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Nagz API
 * Family-oriented AI-mediated nagging system
 * OpenAPI spec version: 0.2.0
 */
import type { PolicyResponseConstraints } from './policyResponseConstraints';
import type { StrategyTemplate } from './strategyTemplate';

export interface PolicyResponse {
  id: string;
  family_id: string;
  owners: unknown[];
  strategy_template: StrategyTemplate;
  constraints: PolicyResponseConstraints;
  status: string;
}